fix: export promptOptions on a subpath
