+++ packages/react-ui/package.json
@@ -46,6 +46,16 @@
         "default": "./dist/genui-lib/index.cjs"
       }
     },
+    "./genui-lib/prompt-options": {
+      "import": {
+        "types": "./dist/genui-lib/prompt-options/index.d.mts",
+        "default": "./dist/genui-lib/prompt-options/index.mjs"
+      },
+      "require": {
+        "types": "./dist/genui-lib/prompt-options/index.d.cts",
+        "default": "./dist/genui-lib/prompt-options/index.cjs"
+      }
+    },
     "./scssUtils": {
       "default": "./dist/styles/cssUtils.scss"
     },
