+++ packages/react-ui/src/genui-lib/openuiLibrary.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import type { ComponentGroup, PromptOptions } from "@openuidev/react-lang";
+import type { ComponentGroup } from "@openuidev/react-lang";
 import { createLibrary } from "@openuidev/react-lang";
 
 // Content
@@ -201,65 +201,8 @@ export const openuiComponentGroups: ComponentGroup[] = [
   },
 ];
 
-// ── Examples ──
-
-export const openuiExamples: string[] = [
-  `Example 1 — Table (column-oriented):
-
-root = Stack([title, tbl])
-title = TextContent("Top Languages", "large-heavy")
-tbl = Table([Col("Language", langs), Col("Users (M)", users), Col("Year", years)])
-langs = ["Python", "JavaScript", "Java", "TypeScript", "Go"]
-users = [15.7, 14.2, 12.1, 8.5, 5.2]
-years = [1991, 1995, 1995, 2012, 2009]`,
-
-  `Example 2 — Bar chart:
-
-root = Stack([title, chart])
-title = TextContent("Q4 Revenue", "large-heavy")
-chart = BarChart(labels, [s1, s2], "grouped")
-labels = ["Oct", "Nov", "Dec"]
-s1 = Series("Product A", [120, 150, 180])
-s2 = Series("Product B", [90, 110, 140])`,
-
-  `Example 3 — Form with validation:
-
-root = Stack([title, form])
-title = TextContent("Contact Us", "large-heavy")
-form = Form("contact", btns, [nameField, emailField, countryField, msgField])
-nameField = FormControl("Name", Input("name", "Your name", "text", { required: true, minLength: 2 }))
-emailField = FormControl("Email", Input("email", "you@example.com", "email", { required: true, email: true }))
-countryField = FormControl("Country", Select("country", countryOpts, "Select...", { required: true }))
-msgField = FormControl("Message", TextArea("message", "Tell us more...", 4, { required: true, minLength: 10 }))
-countryOpts = [SelectItem("us", "United States"), SelectItem("uk", "United Kingdom"), SelectItem("de", "Germany")]
-btns = Buttons([Button("Submit", Action([@ToAssistant("Submit")]), "primary"), Button("Cancel", Action([@ToAssistant("Cancel")]), "secondary")])`,
-
-  `Example 4 — Tabs with mixed content:
-
-root = Stack([title, tabs])
-title = TextContent("React vs Vue", "large-heavy")
-tabs = Tabs([tabReact, tabVue])
-tabReact = TabItem("react", "React", reactContent)
-tabVue = TabItem("vue", "Vue", vueContent)
-reactContent = [TextContent("React is a library by Meta for building UIs."), Callout("info", "Note", "React uses JSX syntax.")]
-vueContent = [TextContent("Vue is a progressive framework by Evan You."), Callout("success", "Tip", "Vue has a gentle learning curve.")]`,
-];
-
-export const openuiAdditionalRules: string[] = [
-  'For grid-like layouts, use Stack with direction "row" and wrap=true. Avoid justify="between" unless you specifically want large gutters.',
-  "For forms, define one FormControl reference per field so controls can stream progressively.",
-  "For forms, always provide the second Form argument with Buttons(...) actions: Form(name, buttons, fields).",
-  "Never nest Form inside Form.",
-  'Use @Reset($var1, $var2) after form submit to restore defaults — not @Set($var, "")',
-  "Multi-query refresh: Action([@Run(mutation), @Run(query1), @Run(query2), @Reset(...)])",
-  "$variables are reactive: changing via Select or @Set re-evaluates all Queries and expressions referencing them",
-  "Use existing components (Tabs, Accordion, Modal) before inventing ternary show/hide patterns",
-];
-
-export const openuiPromptOptions: PromptOptions = {
-  examples: openuiExamples,
-  additionalRules: openuiAdditionalRules,
-};
+// ── Examples & prompt options ──
+export { openuiAdditionalRules, openuiExamples, openuiPromptOptions } from "./prompt-options";
 
 // ── Library ──
